+++ ft_strlcat.c
@@ -6,14 +6,15 @@
 /*   By: kjurkows <kjurkows@student.42warsaw.pl>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/16 12:58:06 by kjurkows          #+#    #+#             */
-/*   Updated: 2026/06/16 13:05:27 by kjurkows         ###   ########.fr       */
+/*   Updated: 2026/06/16 14:22:58 by kjurkows         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlcat(char *dst, const char *src, unsigned int size)
+long unsigned int	ft_strlcat(char *dst, const char *src,
+	long unsigned int size)
 {
-	unsigned int	i;
-	unsigned int	j;
+	long unsigned int	i;
+	long unsigned int	j;
 
 	i = 0;
 	while (i < size - 1 && dst[i])
