+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: kjurkows <kjurkows@student.42warsaw.pl>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/15 20:39:46 by kjurkows          #+#    #+#             */
-/*   Updated: 2026/06/15 20:40:18 by kjurkows         ###   ########.fr       */
+/*   Updated: 2026/06/16 14:22:29 by kjurkows         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlen(const char *s)
+long unsigned int	ft_strlen(const char *s)
 {
-	unsigned int	len;
+	long unsigned int	len;
 
 	len = 0;
 	while (s[len])
